# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.

+++ src/System/System.h
@@ -74,7 +74,7 @@ namespace System {
 
     void initDebug();
 
-    std::string f_to_string(const float value, const int n = 2);
+    std::string f_to_string(const double value, const int n = 2);
 
     sf::Text createDebugString(const std::string &alias);
